rename task to autotask
