+++ gradsflow/autotasks/autoclassification/text/text.py
@@ -26,7 +26,7 @@
 
 import torch
 
-from gradsflow.tasks.engine.autoclassifier import AutoClassifier
+from gradsflow.autotasks.engine.autoclassifier import AutoClassifier
 
 
 # noinspection PyTypeChecker
